# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCloudTouch
  description: Open-Source replacement for discontinued streaming device cloud features
  version: dev-6b24c3b
paths:
...
  /v1/systems/devices/{device_id}:
    get:
      tags:
      - marge
      summary: Get Full Account
      description: "Get full account sync for device.\n\nThis endpoint is called by\
        \ SoundTouch devices on boot to sync:\n- Presets (6 buttons)\n- Recents (recently\
        \ played)\n- Sources (available sources)\n\nArgs:\n    device_id: Device MAC\
        \ address (e.g., \"689E194F7D2F\")\n    marge: Marge service dependency\n\n\
        Returns:\n    XML Response with <boseAccount> structure"
      operationId: get_full_account_v1_systems_devices__device_id__get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/systems/devices/{device_id}/presets:
    get:
      tags:
      - marge
      summary: Get Presets
      description: "Get presets for device.\n\nArgs:\n    device_id: Device MAC address\n\
        \    marge: Marge service dependency\n\nReturns:\n    XML Response with <presets>\
        \ structure"
      operationId: get_presets_v1_systems_devices__device_id__presets_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/systems/devices/{device_id}/recents:
    get:
      tags:
      - marge
      summary: Get Recents
      description: "Get recently played items for device.\n\nArgs:\n    device_id:\
        \ Device MAC address\n    marge: Marge service dependency\n\nReturns:\n  \
        \  XML Response with <recents> structure"
      operationId: get_recents_v1_systems_devices__device_id__recents_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/systems/devices/{device_id}/sources:
    get:
      tags:
      - marge
      summary: Get Sources
      description: "Get available sources for device.\n\nArgs:\n    device_id: Device\
        \ MAC address\n\nReturns:\n    XML Response with <sources> structure"
      operationId: get_sources_v1_systems_devices__device_id__sources_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/systems/devices/{device_id}/devices:
    get:
      tags:
      - marge
      summary: Get Devices
      description: "Get multiroom devices for device.\n\nArgs:\n    device_id: Device\
        \ MAC address\n\nReturns:\n    XML Response with <devices> structure"
      operationId: get_devices_v1_systems_devices__device_id__devices_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/systems/devices/{device_id}/power_on:
    put:
      tags:
      - marge
      summary: Power On
      description: "Device boot notification.\n\nSoundTouch devices call this on power-on\
        \ to notify the server.\n\nArgs:\n    device_id: Device MAC address\n\nReturns:\n\
        \    204 No Content (acknowledgement)"
      operationId: power_on_v1_systems_devices__device_id__power_on_put
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - marge
      summary: Power On
      description: "Device boot notification.\n\nSoundTouch devices call this on power-on\
        \ to notify the server.\n\nArgs:\n    device_id: Device MAC address\n\nReturns:\n\
        \    204 No Content (acknowledgement)"
      operationId: power_on_v1_systems_devices__device_id__power_on_post
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/systems/devices/{device_id}/sourceproviders:
    get:
      tags:
      - marge
      summary: Get Sourceproviders
      description: "Get available source providers for device.\n\nArgs:\n    device_id:\
        \ Device MAC address\n\nReturns:\n    XML Response with <sourceproviders>\
        \ structure"
      operationId: get_sourceproviders_v1_systems_devices__device_id__sourceproviders_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /streaming/account/{account_id}/full:
    get:
      tags:
      - marge
      summary: Streaming Full Account
      description: "Get full account sync via streaming endpoint.\n\nThis is the streaming.bose.com\
        \ version of the account sync endpoint.\nResolves the device_id from the account_id\
        \ (margeAccountUUID) stored\nduring device discovery, then returns that device's\
        \ presets.\n\nArgs:\n    account_id: Account ID (e.g., \"3784726\")\n    marge:\
        \ Marge service dependency\n\nReturns:\n    XML Response with <account> structure"
      operationId: streaming_full_account_streaming_account__account_id__full_get
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scmudc/{device_id}:
    post:
      tags:
      - marge
      summary: Scmudc Reporting
      description: "Device reporting/telemetry endpoint.\n\nDevices periodically call\
        \ this to report status/telemetry data.\nWe acknowledge but don't process\
        \ the data.\n\nArgs:\n    device_id: Device MAC address\n\nReturns:\n    200\
        \ OK"
      operationId: scmudc_reporting_v1_scmudc__device_id__post
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /streaming/device/{device_id}/streaming_token:
    get:
      tags:
      - marge
      summary: Streaming Token
      description: "Return a dummy streaming token for device authentication.\n\n\
        SoundTouch firmware 27.x requests this token before playing a preset.\nIf\
        \ the request fails, the device aborts playback with INVALID_SOURCE.\nThe\
        \ token value is not validated — any non-empty response suffices.\n\nArgs:\n\
        \    device_id: Device MAC address\n\nReturns:\n    JSON with a dummy access\
        \ token"
      operationId: streaming_token_streaming_device__device_id__streaming_token_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/blacklist/{device_id}:
    post:
      tags:
      - marge
      summary: Blacklist Check
      description: "Content blacklist check — always returns empty (nothing blacklisted).\n\
        \nThe device checks this before playing content. If the endpoint is\nunreachable,\
        \ some firmware versions refuse playback.\n\nArgs:\n    device_id: Device\
        \ MAC address\n\nReturns:\n    JSON with empty blacklist"
      operationId: blacklist_check_v1_blacklist__device_id__post
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - marge
      summary: Blacklist Check
      description: "Content blacklist check — always returns empty (nothing blacklisted).\n\
        \nThe device checks this before playing content. If the endpoint is\nunreachable,\
        \ some firmware versions refuse playback.\n\nArgs:\n    device_id: Device\
        \ MAC address\n\nReturns:\n    JSON with empty blacklist"
      operationId: blacklist_check_v1_blacklist__device_id__get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...